backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.3.20
langchain-openai==0.3.7
python-dotenv==1.0.1
